Rename clientapp to client_app
