+++ src/py/flwr/client/__init__.py
@@ -19,7 +19,7 @@ from .app import run_client_app as run_client_app
 from .app import start_client as start_client
 from .app import start_numpy_client as start_numpy_client
 from .client import Client as Client
-from .clientapp import ClientApp as ClientApp
+from .client_app import ClientApp as ClientApp
 from .numpy_client import NumPyClient as NumPyClient
 from .typing import ClientFn as ClientFn
