+++ src/py/flwr/client/app.py
@@ -23,7 +23,7 @@ from pathlib import Path
 from typing import Callable, ContextManager, Optional, Tuple, Union
 
 from flwr.client.client import Client
-from flwr.client.clientapp import ClientApp
+from flwr.client.client_app import ClientApp
 from flwr.client.typing import ClientFn
 from flwr.common import GRPC_MAX_MESSAGE_LENGTH, EventType, Message, event
 from flwr.common.address import parse_address
@@ -36,7 +36,7 @@ from flwr.common.constant import (
 )
 from flwr.common.logger import log, warn_deprecated_feature, warn_experimental_feature
 
-from .clientapp import load_client_app
+from .client_app import load_client_app
 from .grpc_client.connection import grpc_connection
 from .grpc_rere_client.connection import grpc_request_response
 from .message_handler.message_handler import handle_control_message
